+++ apps/meteor/app/slackbridge/server/SlackAPI.ts
@@ -11,7 +11,9 @@ export class SlackAPI {
 
 	async getChannels(cursor = null) {
 		let channels = [];
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/conversations.list', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -26,7 +28,7 @@ export class SlackAPI {
 
 		if (response && response && Array.isArray(response.channels) && response.channels.length > 0) {
 			channels = channels.concat(response.channels);
-			if (response.response_metadata && response.response_metadata.next_cursor) {
+			if (response.response_metadata?.next_cursor) {
 				const nextChannels = await this.getChannels(response.response_metadata.next_cursor);
 				channels = channels.concat(nextChannels);
 			}
@@ -37,7 +39,9 @@ export class SlackAPI {
 
 	async getGroups(cursor = null) {
 		let groups = [];
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/conversations.list', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -52,7 +56,7 @@ export class SlackAPI {
 
 		if (response && response && Array.isArray(response.channels) && response.channels.length > 0) {
 			groups = groups.concat(response.channels);
-			if (response.response_metadata && response.response_metadata.next_cursor) {
+			if (response.response_metadata?.next_cursor) {
 				const nextGroups = await this.getGroups(response.response_metadata.next_cursor);
 				groups = groups.concat(nextGroups);
 			}
@@ -62,7 +66,9 @@ export class SlackAPI {
 	}
 
 	async getRoomInfo(roomId) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch(`https://slack.com/api/conversations.info`, {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -81,8 +87,9 @@ export class SlackAPI {
 		let members = [];
 		let currentCursor = '';
 		for (let index = 0; index < num_members; index += MAX_MEMBERS_PER_CALL) {
-			// eslint-disable-next-line no-await-in-loop
 			const request = await fetch('https://slack.com/api/conversations.members', {
+				// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
+				ignoreSsrfValidation: true,
 				headers: {
 					Authorization: `Bearer ${this.token}`,
 				},
@@ -92,11 +99,10 @@ export class SlackAPI {
 					...(currentCursor && { cursor: currentCursor }),
 				},
 			});
-			// eslint-disable-next-line no-await-in-loop
 			const response = await request.json();
 			if (response && response && request.status === 200 && request.ok && Array.isArray(response.members)) {
 				members = members.concat(response.members);
-				const hasMoreItems = response.response_metadata && response.response_metadata.next_cursor;
+				const hasMoreItems = response.response_metadata?.next_cursor;
 				if (hasMoreItems) {
 					currentCursor = response.response_metadata.next_cursor;
 				}
@@ -106,7 +112,9 @@ export class SlackAPI {
 	}
 
 	async react(data) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/reactions.add', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -118,7 +126,9 @@ export class SlackAPI {
 	}
 
 	async removeReaction(data) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/reactions.remove', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -130,7 +140,9 @@ export class SlackAPI {
 	}
 
 	async removeMessage(data) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/chat.delete', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -142,7 +154,9 @@ export class SlackAPI {
 	}
 
 	async sendMessage(data) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/chat.postMessage', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -153,7 +167,9 @@ export class SlackAPI {
 	}
 
 	async updateMessage(data) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/chat.update', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -165,7 +181,9 @@ export class SlackAPI {
 	}
 
 	async getHistory(options) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch(`https://slack.com/api/conversations.history`, {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -176,7 +194,9 @@ export class SlackAPI {
 	}
 
 	async getPins(channelId) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/pins.list', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -189,7 +209,9 @@ export class SlackAPI {
 	}
 
 	async getUser(userId) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/users.info', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${this.token}`,
 			},
@@ -202,7 +224,9 @@ export class SlackAPI {
 	}
 
 	static async verifyToken(token) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/auth.test', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${token}`,
 			},
@@ -213,7 +237,9 @@ export class SlackAPI {
 	}
 
 	static async verifyAppCredentials({ botToken, appToken }) {
+		// SECURITY: the URL is a default hardcoded value or an envvar/setting set by an admin. It's safe to disable this check.
 		const request = await fetch('https://slack.com/api/apps.connections.open', {
+			ignoreSsrfValidation: true,
 			headers: {
 				Authorization: `Bearer ${appToken}`,
 			},