+++ apps/uikit-playground/src/Components/FlowContainer/ControlButtons/index.ts
@@ -1,1 +1,1 @@
-export {default} from './ControlButtons';
+export { default } from './ControlButtons';